Order. moved model to entities,added Service and deleted some unused imports
