+++ src/main/java/org/newton/webshop/models/entities/Order.java
@@ -1,9 +1,10 @@
-package org.newton.webshop.models;
+package org.newton.webshop.models.entities;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import org.hibernate.annotations.GenericGenerator;
+import org.newton.webshop.models.Cart;
 
 import javax.persistence.*;
 import java.time.LocalDateTime;
