+++ src/main/java/org/newton/webshop/repositories/OrderRepository.java
@@ -1,7 +1,7 @@
 package org.newton.webshop.repositories;
 
-import org.newton.webshop.models.Order;
-import org.springframework.data.jpa.repository.JpaRepository;
+import org.newton.webshop.models.entities.Order;
+import org.springframework.data.repository.CrudRepository;
 
-public interface OrderRepository extends JpaRepository<Order, String> {
+public interface OrderRepository extends CrudRepository<Order, String> {
 }
